+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Drop support for Windows 7 and 8
 - Drop support for macOS 10.13 and 10.14
 - Add playback history to queue panel
+- Add queue slide-out transition
 - Fix dragging tracks to up next/autoplay queue boundary
 - Fix button edge not clickable due to zoom
 - Update dependencies